added error messages to login page and fixed add user conditional check
login page will now display a red server response message indicating what is wrong if they try to create an existing username or login with the wrong password.  Also, there was an error in the add_user function because it would disallow a person from creating any user at all rather than only duplicate ones.  This was fixed in this commit.