customized database functions to allow direct message input
added keyword arguments to channelid and made it optional.  Enabled write_message function in wsgi.py to executed based on the message_type parameter passed to it in the header.  Now the write_message function will either write  message with a channel ID for  a channel message or write a message with a recipient_user_id based on the received username.