user select dropdown for DMs is working.
the problem was being called when I fetched a list from the server and stored it in session storage.  I did not realize session storage doesnt support arrays so when I went to map the results my javascrip was trying to map a string.  Solved by converting in and out of JSON.