fixed doublelogin bug
app.js starts by setting startinglogin to false and then updates to true only if loginstatus is true in session storage.  However, upon sucessful login, the setloginstatus was udpated to true before sessionstorage was updated.  This triggered a rerender of App.js PRIOR to sessionstorage loginstatus being set to true.  So the result was you would be logged in but redirected to the login page.  I switched the order in login.js so that the sessionstorage was updated BEFORE the stateupdate/rerender trigger.  Problem fixed.